refactor: httpUtil에 몰려있던 파싱 기능을 별도로 분리
